+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -198,18 +198,13 @@ object HybrisIcons {
         val GROUP_MAP = getIcon("/icons/typeSystem/groupByMap.svg")
         val GROUP_RELATION = getIcon("/icons/typeSystem/groupByRelation.svg")
         val DESCRIPTION = AllIcons.Windows.Help
-        val ENUM = getIcon("/icons/typeSystem/enum.svg")
         val ENUM_VALUE = getIcon("/icons/typeSystem/enumValue.svg")
-        val ATOMIC = getIcon("/icons/typeSystem/atomic.svg")
         val ITEM = getIcon("/icons/typeSystem/item.svg")
-        val MAP = getIcon("/icons/typeSystem/map.svg")
-        val RELATION = getIcon("/icons/typeSystem/relation.svg")
         val CUSTOM_PROPERTY = getIcon("/icons/typeSystem/customProperty.svg")
         val ATTRIBUTE = getIcon("/icons/typeSystem/attribute.svg")
         val ORDERING_ATTRIBUTE = getIcon("/icons/typeSystem/orderingAttribute.svg")
         val RELATION_SOURCE = getIcon("/icons/typeSystem/relationSource.svg")
         val RELATION_TARGET = getIcon("/icons/typeSystem/relationTarget.svg")
-        val COLLECTION = getIcon("/icons/typeSystem/collection.svg")
         val INDEX = getIcon("/icons/typeSystem/index.svg")
         val INDEX_UNIQUE = getIcon("/icons/typeSystem/indexUnique.svg")
         val INDEX_REPLACE = getIcon("/icons/typeSystem/indexReplace.svg")
@@ -217,6 +212,15 @@ object HybrisIcons {
         val HEADER_ABBREVIATION = getIcon("/icons/typeSystem/headerAbbreviation.svg")
         val ALTERNATIVE_DECLARATION = AllIcons.Actions.Forward
         val SIBLING = AllIcons.Gutter.OverridenMethod
+
+        object Types {
+            val ATOMIC = getIcon("/icons/typeSystem/atomic.svg")
+            val MAP = getIcon("/icons/typeSystem/map.svg")
+            val RELATION = getIcon("/icons/typeSystem/relation.svg")
+            val ENUM = getIcon("/icons/typeSystem/enum.svg")
+            val COLLECTION = getIcon("/icons/typeSystem/collection.svg")
+            val PRIMITIVE = HybrisIcons.Types.PRIMITIVE
+        }
 
         object Preview {
             object Actions {