+++ src/com/intellij/idea/plugin/hybris/impex/psi/references/ImpexFunctionTSAttributeReference.kt
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -58,7 +58,7 @@ class ImpexFunctionTSAttributeReference(owner: ImpexParameter) : TSReferenceBase
         // if an inline type already present, we should not suggest any other types
         if (element.inlineTypeName != null) return emptyArray()
         return TSCompletionService.getInstance(element.project)
-            .getImpexInlineTypeCompletions(element.project, element)
+            .getImpExInlineTypeCompletions(element.project, element)
             .toTypedArray()
     }