+++ src/com/intellij/idea/plugin/hybris/system/type/model/deployment/AttributeMapping.java
@@ -23,30 +23,12 @@ package com.intellij.idea.plugin.hybris.system.type.model.deployment;
 
 import com.intellij.util.xml.DomElement;
 import com.intellij.util.xml.GenericAttributeValue;
-import com.intellij.util.xml.Required;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * null:attribute-mappingElemType interface.
  */
 public interface AttributeMapping extends DomElement {
-
-    /**
-     * Returns the value of the simple content.
-     *
-     * @return the value of the simple content.
-     */
-    @NotNull
-    @Required
-    String getValue();
-
-    /**
-     * Sets the value of the simple content.
-     *
-     * @param value the new value to set
-     */
-    void setValue(@NotNull String value);
-
 
     /**
      * Returns the value of the persistence-name child.