+++ src/com/intellij/idea/plugin/hybris/system/type/model/deployment/Table.java
@@ -21,32 +21,15 @@
 
 package com.intellij.idea.plugin.hybris.system.type.model.deployment;
 
+import com.intellij.util.xml.Attribute;
 import com.intellij.util.xml.DomElement;
 import com.intellij.util.xml.GenericAttributeValue;
-import com.intellij.util.xml.Required;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * null:tableElemType interface.
  */
 public interface Table extends DomElement {
-
-    /**
-     * Returns the value of the simple content.
-     *
-     * @return the value of the simple content.
-     */
-    @NotNull
-    @Required
-    String getValue();
-
-    /**
-     * Sets the value of the simple content.
-     *
-     * @param value the new value to set
-     */
-    void setValue(@NotNull String value);
-
 
     /**
      * Returns the value of the name child.